refactor: AccountViewのリファクタリングをした
