+++ Gitagram/View/AccountView/AccountView.swift
@@ -1,0 +1,52 @@
+//
+//  AccountQRView.swift
+//  Gitagram
+//
+//  Created by 伊藤璃乃 on 2024/05/03.
+//
+
+import SwiftUI
+import CoreImage.CIFilterBuiltins
+
+struct AccountView: View {
+    private static let cardCornerRadius: CGFloat = 20
+    private static let cardSize: CGSize = .init(width: 367, height: 512)
+    private let viewModel = AccountViewModel()
+    
+    @State private var isSharing: Bool = false
+    @State var developer: Developer
+    
+    var body: some View {
+        NavigationView {
+            ZStack {
+                ZStack {
+                    KiraKiraView()
+                    RoundedRectangle(cornerRadius: Self.cardCornerRadius)
+                        .stroke(Color.white, lineWidth: 1)
+                }
+                .frame(width: Self.cardSize.width, height: Self.cardSize.height)
+                .shadow(color: Color.black.opacity(0.5), radius: 10, x: 0, y: 20)
+                
+                QRView(inputText: developer.gitHubURL)
+                    .padding()
+                
+                RichText(inputText: developer.name)
+            }
+            .padding(50)
+            .toolbar {
+                ToolbarItem(placement: .navigationBarTrailing) {
+                    Button(action: {
+                        isSharing.toggle()
+                    }) {
+                        Image(systemName: "square.and.arrow.up")
+                    }
+                }
+            }
+            .sheet(isPresented: $isSharing, content: {
+                ShareSheet(activityItems: [viewModel.captureViewAsImage()])
+            })
+        }
+    }
+}
+
+
