+++ Gitagram/View/AccountView/Component/StrokeBackground.swift
